+++ lib/features/account/boardgames/widgets/bg_info_card.dart
