+++ lib/features/account/mechanics/widgets/mechanic_dialog.dart
