+++ lib/features/account/mechanics/widgets/show_mechs/widgets/dismissible_mech.dart
